test suite on KIC9574283
provided a new python function (in external/compare_in2realspectrum.py) that allows you to compare a real spectrum (saved in a *.sav file) with the outputs of an *.in file

The main.cfg.test.KIC9574283 gives you a configuration file for comparing the simulation with KIC95774283, as it was analysed in Deheuvels et al. 2014 (https://arxiv.org/pdf/1401.3096.pdf)

A spectrum of that star is also provided for test purpose.