minor update
the solver had some crash for stars with low Dnu. This is due to the fix choice of the grid density to find solutions. Now in case of failure, the code attempt to increase the grid density iteratively (up to 4 times higher than the initial density) until really showing an error message.
This allows to explore a lower range of Dnu.