v1.2 : new model generate_cfg_from_synthese_file_Wscaled_aj_GRANscaled
New model for main sequence stars that allow you to rescale the noise background following the scaling of the noise with numax as described by Karoff et al. 2010 (or Kallinger et al. 2014).
This is basically an extension to the main sequence of the same capability already implemented in red giant spectrum models.